Добавляет trip-presenter
